+++ packages/contracts/contracts/libs/OperationUtils.sol
@@ -34,19 +34,24 @@ library OperationUtils {
             uint256 numJoinSplitsForOp = op.joinSplits.length;
             for (uint256 j = 0; j < numJoinSplitsForOp; j++) {
                 proofs[index] = op.joinSplits[j].proof;
-                allPis[index] = new uint256[](9);
+                allPis[index] = new uint256[](11);
                 allPis[index][0] = op.joinSplits[j].newNoteACommitment;
                 allPis[index][1] = op.joinSplits[j].newNoteBCommitment;
                 allPis[index][2] = op.joinSplits[j].commitmentTreeRoot;
                 allPis[index][3] = op.joinSplits[j].publicSpend;
                 allPis[index][4] = op.joinSplits[j].nullifierA;
                 allPis[index][5] = op.joinSplits[j].nullifierB;
-                allPis[index][6] = digests[i];
-                allPis[index][7] = op
+                allPis[index][6] = op.joinSplits[j].encSenderCanonAddrC1X;
+                allPis[index][7] = op.joinSplits[j].encSenderCanonAddrC2X;
+                allPis[index][8] = digests[i];
+                allPis[index][9] = op
                     .joinSplits[j]
                     .encodedAsset
                     .encodedAssetAddr;
-                allPis[index][8] = op.joinSplits[j].encodedAsset.encodedAssetId;
+                allPis[index][10] = op
+                    .joinSplits[j]
+                    .encodedAsset
+                    .encodedAssetId;
                 index++;
             }
         }
@@ -115,12 +120,13 @@ library OperationUtils {
     }
 
     // From https://ethereum.stackexchange.com/questions/83528
+    // returns empty string if no revert message
     function getRevertMsg(
         bytes memory reason
     ) internal pure returns (string memory) {
         // If the _res length is less than 68, then the transaction failed silently (without a revert message)
         if (reason.length < 68) {
-            return "Transaction reverted silently";
+            return "";
         }
 
         assembly {