+++ packages/local-prover/scripts/genJoinSplitTestCase.ts
@@ -10,6 +10,7 @@ import {
   MerkleProofInput,
   EncodedNote,
   StealthAddressTrait,
+  randomFr,
 } from "@nocturne-xyz/sdk";
 
 const ROOT_DIR = findWorkspaceRoot()!;
@@ -149,12 +150,16 @@ const joinsplitInputs: JoinSplitInputs = {
   newNoteB,
   merkleProofA: merkleProofAInput,
   merkleProofB: merkleProofBInput,
+  encRandomness: randomFr(),
 };
 console.log(joinsplitInputs);
 
 (async () => {
   const prover = new WasmJoinSplitProver(WASM_PATH, ZKEY_PATH, VKEY);
+  const startTime = Date.now();
   const proof = await prover.proveJoinSplit(joinsplitInputs);
+  console.log("Proof generated in: ", Date.now() - startTime, "ms");
+
   if (!(await prover.verifyJoinSplitProof(proof))) {
     throw new Error("Proof invalid!");
   }